Added links

+++ tools/publish-pages.ps1
@@ -9,12 +9,24 @@ Write-Host "Source File: $sourceFile"
 Write-Host "Destination Dir: $destinationDir"
 Write-Host "Urls File: $urlsFile"
 
-foreach ($i in $urls.data)
+$userInput = Read-Host -Prompt "All folders and files will be deleted in $destinationDir. Enter Y to confirm or any other key to skip: "
+if($userInput -eq "Y")
 {
-    $destinationSubDir = Join-Path -Path $destinationDir -ChildPath $i
-    Write-Host "Destination Sub Dir:  $destinationSubDir"
-    if (!(Test-Path -Path $destinationSubDir)) {New-Item $destinationSubDir -Type Directory}
-    Copy-Item $sourceFile -Destination $destinationSubDir
+    Remove-Item -Path $destinationDir -Recurse -Force
+    New-Item $destinationDir -Type Directory
+}
+
+$userInput = Read-Host -Prompt "Would you like to continue publishing. Enter Y to confirm or any other key to skip: "
+
+if($userInput -eq "Y")
+{
+    foreach ($i in $urls.data)
+    {
+        $destinationSubDir = Join-Path -Path $destinationDir -ChildPath $i
+        Write-Host "Destination Sub Dir:  $destinationSubDir"
+        if (!(Test-Path -Path $destinationSubDir)) {New-Item $destinationSubDir -Type Directory}
+        Copy-Item $sourceFile -Destination $destinationSubDir
+    }
 }
 
 Read-Host -Prompt "Press any key to exit..."
